We made a Pseudocode code last time:
if (ColorSenseOut =< a : 29 | r : 89 | g : 56 | b : 2a) {
[make robot stop];
} else {
[keep moving];
}

Now, we finally coded the actual code out and it works.
We copied some of the telemetry from SensorColor to tell us what the colors it senses are. Now that the robot can stop when it senses color, we made our own telemetry to tell us when it senses that color and finishes.