I did add things and delete things. But it did get me further in progress.
Example: state = SamplingStateMachine.RobotState.CheckScreen;

I changed the state it would go to from PushBloock to CheckScreen. This did fix the problem that I said from last commit, but it also extended the time it was running. If I can do a few tweaks, maybe this problem will be fixed entirely.